Added flush_cache_and_wait() in prep for future change to cause flush before getattr
